+++ cascalog-core/src/java/cascalog/ClojureBuffer.java
@@ -27,6 +27,7 @@ import clojure.lang.IFn;
 import clojure.lang.ISeq;
 import clojure.lang.IteratorSeq;
 import clojure.lang.RT;
+import clojure.lang.Var;
 
 public class ClojureBuffer extends ClojureCascadingBase implements Buffer {
 
@@ -35,13 +36,18 @@ public class ClojureBuffer extends ClojureCascadingBase implements Buffer {
   }
 
   public void operate(FlowProcess flow_process, BufferCall call) {
-    ISeq resultSeq = RT.seq(invokeFunction(IteratorSeq
-        .create(new TupleSeqConverter(call.getArgumentsIterator()))));
-    TupleEntryCollector collector = call.getOutputCollector();
-    while (resultSeq != null) {
-      Object obj = resultSeq.first();
-      collector.add(Util.coerceToTuple(obj));
-      resultSeq = resultSeq.next();
+    Var.pushThreadBindings(bindingMap);
+    try {
+      ISeq resultSeq = RT.seq(invokeFunction(IteratorSeq
+          .create(new TupleSeqConverter(call.getArgumentsIterator()))));
+      TupleEntryCollector collector = call.getOutputCollector();
+      while (resultSeq != null) {
+        Object obj = resultSeq.first();
+        collector.add(Util.coerceToTuple(obj));
+        resultSeq = resultSeq.next();
+      }
+    } finally {
+      Var.popThreadBindings();
     }
   }
 }