+++ cascalog-core/src/java/cascalog/ClojureMap.java
@@ -1,14 +1,14 @@
 /*
     Copyright 2010 Nathan Marz
- 
-    Project and contact information: http://www.cascalog.org/ 
+
+    Project and contact information: http://www.cascalog.org/
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
     You may obtain a copy of the License at
-   
+
         http://www.apache.org/licenses/LICENSE-2.0
-   
+
     Unless required by applicable law or agreed to in writing, software
     distributed under the License is distributed on an "AS IS" BASIS,
     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -24,15 +24,21 @@ import cascading.operation.FunctionCall;
 import cascading.tuple.Fields;
 import clojure.lang.IFn;
 import clojure.lang.ISeq;
+import clojure.lang.Var;
 
 public class ClojureMap extends ClojureCascadingBase implements Function {
   public ClojureMap(Fields outputFields, IFn fn) {
     super(outputFields, fn);
   }
 
-  public void operate(FlowProcess fp, FunctionCall call) {
-    ISeq fnArgs = Util.coerceFromTuple(call.getArguments().getTuple());
-    Object res = applyFunction(fnArgs);
-    call.getOutputCollector().add(Util.coerceToTuple(res));
+  public void operate(FlowProcess fp, final FunctionCall call) {
+    Var.pushThreadBindings(bindingMap);
+    try {
+      ISeq fnArgs = Util.coerceFromTuple(call.getArguments().getTuple());
+      Object res = applyFunction(fnArgs);
+      call.getOutputCollector().add(Util.coerceToTuple(res));
+    } finally {
+      Var.popThreadBindings();
+    }
   }
 }